+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 # selenium-components-0 (alpha)
 
+## selenium-components-0.6.8
+
+- Fast access for select options, that "contain" a value
+
 ## selenium-components-0.6.7
 
 - Fast access for radio components by index